@@ -14,8 +14,8 @@
     "lodash": "^4.17.5",
     "moment": "^2.21.0",
     "prop-types": "^15.6.1",
-    "react": "^16.2.0",
-    "react-dom": "^16.2.0",
+    "react": "^16.4.1",
+    "react-dom": "^16.4.1",
     "react-router-dom": "^4.2.2",
     "react-scripts": "1.1.1",
     "request": "^2.83.0",
